+++ doc/source/releasenotes.rst
@@ -7,6 +7,8 @@ Release notes
 1.0.1
 ======
 * Added the ``ignore_sizes`` keyword to :py:func:`~clease.plot_post_process.plot_eci`
+* Changing the maximum cluster diameter will now clear any cached clusters, and
+  requires a new build.
 
 
 1.0.0